+++ js/data.js
@@ -1,0 +1,40 @@
+//Тексты комментариев
+const AUTHOR_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+//Имена авторов комментариев
+const COMMENT_AUTHOR_NAME = [
+  'Коржик',
+  'Карамелька',
+  'Компот',
+  'Сажик',
+  'Лапочка',
+  'Шуруп',
+  'Горчица',
+  'Изюм',
+  'Бантик',
+  'Пончик'
+];
+
+//Номер фото для аватара
+const NUMBER_AVATAR = {
+  MIN: 1,
+  MAX: 6,
+};
+
+//Количество комментариев
+const NUMBER_COMMENT = {
+  MIN: 1,
+  MAX: 10,
+};
+
+// Необходимое количество фотографий
+const REQUIRED_PHOTO_COUNT = 25;
+
+export {AUTHOR_MESSAGES, COMMENT_AUTHOR_NAME, NUMBER_AVATAR, NUMBER_COMMENT, REQUIRED_PHOTO_COUNT};
